requirements.txt
asyncpg
discord.py
jishaku
ruamel.yaml
tabulate
